Rename perf->performance
