+++ src/physicalai/benchmark/performance/__init__.py
@@ -5,9 +5,9 @@
 
 Public API::
 
-    from physicalai.benchmark.perf import InferenceBenchmark
+    from physicalai.benchmark.performance import InferenceBenchmark
 """
 
-from physicalai.benchmark.perf.inference_benchmark import InferenceBenchmark
+from physicalai.benchmark.performance.inference_benchmark import InferenceBenchmark
 
 __all__ = ["InferenceBenchmark"]
